add error boundary
