+++ src/views/Predictions/components/LabelPrice.tsx
@@ -4,6 +4,7 @@ import { Text } from '@pancakeswap/uikit'
 import { formatBigNumberToFixed } from 'utils/formatBalance'
 import styled from 'styled-components'
 import { BigNumber } from '@ethersproject/bignumber'
+import ErrorBoundary from 'components/ErrorBoundary'
 
 const Price = styled(Text)`
   height: 18px;
@@ -23,13 +24,15 @@ const LabelPrice: React.FC<React.PropsWithChildren<LabelPriceProps>> = ({ price 
   const priceAsNumber = useMemo(() => parseFloat(formatBigNumberToFixed(price, 4, 8)), [price])
 
   return (
-    <CountUp start={0} preserveValue delay={0} end={priceAsNumber} prefix="$" decimals={4} duration={1}>
-      {({ countUpRef }) => (
-        <Price fontSize="12px">
-          <span ref={countUpRef} />
-        </Price>
-      )}
-    </CountUp>
+    <ErrorBoundary>
+      <CountUp start={0} preserveValue delay={0} end={priceAsNumber} prefix="$" decimals={4} duration={1}>
+        {({ countUpRef }) => (
+          <Price fontSize="12px">
+            <span ref={countUpRef} />
+          </Price>
+        )}
+      </CountUp>
+    </ErrorBoundary>
   )
 }
