+++ src/views/Predictions/components/RoundCard/LiveRoundPrice.tsx
@@ -3,6 +3,7 @@ import { BigNumber } from '@ethersproject/bignumber'
 import CountUp from 'react-countup'
 import { Skeleton, TooltipText } from '@pancakeswap/uikit'
 import { formatBigNumberToFixed } from 'utils/formatBalance'
+import { ErrorBoundary } from 'components/ErrorBoundary'
 
 interface LiveRoundPriceProps {
   isBull: boolean
@@ -15,13 +16,15 @@ const LiveRoundPrice: React.FC<React.PropsWithChildren<LiveRoundPriceProps>> = (
   const priceColor = isBull ? 'success' : 'failure'
 
   return (
-    <CountUp start={0} preserveValue delay={0} end={priceAsNumber} prefix="$" decimals={4} duration={1}>
-      {({ countUpRef }) => (
-        <TooltipText bold color={priceColor} fontSize="24px" style={{ minHeight: '36px' }}>
-          {price.gt(0) ? <span ref={countUpRef} /> : <Skeleton height="36px" width="94px" />}
-        </TooltipText>
-      )}
-    </CountUp>
+    <ErrorBoundary>
+      <CountUp start={0} preserveValue delay={0} end={priceAsNumber} prefix="$" decimals={4} duration={1}>
+        {({ countUpRef }) => (
+          <TooltipText bold color={priceColor} fontSize="24px" style={{ minHeight: '36px' }}>
+            {price.gt(0) ? <span ref={countUpRef} /> : <Skeleton height="36px" width="94px" />}
+          </TooltipText>
+        )}
+      </CountUp>
+    </ErrorBoundary>
   )
 }
